Define mysql users with their own allowed host
In the current implementation of puphpet, MySQL user accounts are pretty much always created as `someuser@localhost`. This blocks SSH-tunneled connections from the host machine since the use of `localhost` only allows socket connections inside the VM. The bottom line is that using Sequel Pro with a puphpet-based vagrant VM is hard because of the current limitation in specifying allowed hosts for a given mysql user.

This change allows puphpet users to specify a `config.yaml/mysql:databases:xyz:user:` like `someuser@%` or `someuser@192.168.56.%`. This can **not** currently be done using `config.yaml/mysql:databases:xyz:host:` because that value is also used to **connect** to the database server when loading an .sql file for example. One "host" value can not simultaneously represent both:

* What host (that is running mysqld) do you wish to **operate on**? and
* What host do you want this new user to be allowed to **connect from**?

Hence this change. It is backwards compatible in that if no `@` is present in the `user:` string, the `host:` setting will be used (even if semantically it has a different meaning.)